arreglando el registro de apredis

+++ back-end/sprint-boot/src/main/java/com/sena_proyecto_car_2025/service/CertificadosService.java
@@ -2,9 +2,9 @@ package com.sena_proyecto_car_2025.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.sena_proyecto_car_2025.model.Certificados;
 import com.sena_proyecto_car_2025.repository.ICertificados;
+import java.util.Optional;
 
 @Service
 public class CertificadosService {
@@ -12,30 +12,31 @@ public class CertificadosService {
     @Autowired
     private ICertificados certificadosRepository;
 
-    public boolean save(Certificados certificado) {
-        certificadosRepository.save(certificado);
-        return true;
-    }
-
-    // Obtener todos los registros
     public Iterable<Certificados> findAll() {
         return certificadosRepository.findAll();
     }
 
-    // Obtener un registro por ID
     public Certificados findById(Integer id) {
-        return certificadosRepository.findById(id).orElse(null);
+        Optional<Certificados> optionalCertificado = certificadosRepository.findById(id);
+        return optionalCertificado.orElse(null);
     }
 
-    // Actualizar un registro
-    public boolean update(Certificados certificado) {
-        certificadosRepository.save(certificado);
-        return true;
+    public Certificados save(Certificados certificado) {
+        return certificadosRepository.save(certificado);
     }
 
-    // Eliminar un registro
+    public Certificados update(Certificados certificado) {
+        if (certificadosRepository.existsById(certificado.getIdCertificado())) {
+            return certificadosRepository.save(certificado);
+        }
+        return null;
+    }
+
     public boolean delete(Integer id) {
-        certificadosRepository.deleteById(id);
-        return true;
+        if (certificadosRepository.existsById(id)) {
+            certificadosRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
-} 
+}
